+++ src/components/Community/Common/CategoryNavigation/CategoryNavigation.styles.ts
